feat(pam): port the request-detail specs and stories onto the dialog
The page's rendering, decision-log and action-gating coverage moves to the dialog
spec unchanged; the route spec is rewritten for the host, covering what it opens the
dialog with, and the three ways out - dismissed with history, dismissed cold, and
left by a navigation that must not double back.

The detail-service spec carries over untouched.
